system info issue fix
